refactor: use PrefsManager
