+++ packages/scaffold/src/utils/prefs.ts
@@ -76,3 +76,32 @@ export class PrefsManager {
     return this.prefs;
   }
 }
+
+export function renderPluginPrefsDts(prefs: Prefs, prefix: string) {
+  const dtsContent = `
+/* prettier-ignore */
+/* eslint-disable */
+// @ts-nocheck
+// Generated by zotero-plugin-scaffold
+
+type _PluginPrefsMap = {
+  ${Object.entries(prefs).map(([key, value]) => {
+    return `"${key.replace(prefix, "")}": ${typeof value};`;
+  }).join("\n  ")}
+};
+
+type PluginPrefKey<K extends keyof _PluginPrefsMap> = \`${prefix}\${K}\`;
+
+type PluginPrefsMap = {
+  [K in keyof _PluginPrefsMap as PluginPrefKey<K>]: _PluginPrefsMap[K]
+};
+
+declare namespace _ZoteroTypes {
+  interface Prefs {
+    get: <K extends keyof PluginPrefsMap>(key: K, global?: boolean) => PluginPrefsMap[K];
+    set: <K extends keyof PluginPrefsMap>(key: K, value: PluginPrefsMap[K], global?: boolean) => any;
+  }
+}
+`;
+  return dtsContent;
+}
